Cleanup Layered Image Windows support
