+++ substratevm/src/com.oracle.svm.hosted/src/com/oracle/svm/hosted/image/NativeImageViaCC.java
@@ -38,7 +38,6 @@ import java.util.function.Function;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 
-import com.oracle.svm.core.image.ImageHeapLayoutInfo;
 import org.graalvm.nativeimage.Platform;
 import org.graalvm.nativeimage.impl.InternalPlatform;
 
@@ -47,7 +46,7 @@ import com.oracle.svm.core.BuildArtifacts;
 import com.oracle.svm.core.BuildArtifacts.ArtifactType;
 import com.oracle.svm.core.LinkerInvocation;
 import com.oracle.svm.core.SubstrateOptions;
-import com.oracle.svm.shared.util.SubstrateUtil;
+import com.oracle.svm.core.image.ImageHeapLayoutInfo;
 import com.oracle.svm.core.util.InterruptImageBuilding;
 import com.oracle.svm.hosted.DeadlockWatchdog;
 import com.oracle.svm.hosted.FeatureImpl.BeforeImageWriteAccessImpl;
@@ -57,6 +56,7 @@ import com.oracle.svm.hosted.c.util.FileUtils;
 import com.oracle.svm.hosted.meta.HostedMetaAccess;
 import com.oracle.svm.hosted.meta.HostedMethod;
 import com.oracle.svm.hosted.meta.HostedUniverse;
+import com.oracle.svm.shared.util.SubstrateUtil;
 
 import jdk.graal.compiler.debug.DebugContext;
 import jdk.graal.compiler.debug.Indent;
@@ -153,6 +153,9 @@ public abstract class NativeImageViaCC extends NativeImage {
             if (status != 0) {
                 String output = String.join(System.lineSeparator(), lines);
                 throw handleLinkerFailure("Linker command exited with " + status, commandLine, output);
+            }
+            if (pecoffForwardRefs != null) {
+                inv.verifyLinkerOutput(lines, pecoffForwardRefs.getRecordedForwardReferenceSymbols());
             }
 
             Path imagePath = inv.getOutputFile();
